icap-v7-master/ai-agent/requirements.txt
# AI Agent Service - RabbitMQ Worker
# Core dependencies for message processing and LLM integration

# Message Queue
pika==1.3.2

# Redis for caching and pub/sub
redis[hiredis]==5.0.1

# LLM Framework
transformers==4.52.3
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-openai==0.3.18
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
openai==1.82.0
ollama==0.4.8
tiktoken==0.9.0
sentencepiece==0.2.0
tokenizers==0.21.1

# Core ML dependencies
torch==2.7.0
numpy==2.2.6
PyYAML==6.0.2

# Data processing
pydantic==2.11.5
pydantic-settings==2.9.1
python-dotenv==1.1.0
requests==2.32.3
RapidFuzz==3.13.0
python-dateutil==2.8.2

# Image processing (for document analysis)
opencv-python-headless==4.11.0.86
Pillow==10.4.0

# External API clients
groq

# Development tools
